+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change save callbacks from `PresetTimeCallback` to `FunctionCallingCallback`. ([#410])
 - Align `eigenstates` and `eigenenergies` to QuTiP. ([#411])
 - Introduce `vector_to_operator` and `operator_to_vector`. ([#413])
+- Change structure of tolerances between pure and mixed state dynamics. ([#415])
 
 ## [v0.27.0]
 Release date: 2025-02-14
@@ -149,3 +150,4 @@ Release date: 2024-11-13
 [#410]: https://github.com/qutip/QuantumToolbox.jl/issues/410
 [#411]: https://github.com/qutip/QuantumToolbox.jl/issues/411
 [#413]: https://github.com/qutip/QuantumToolbox.jl/issues/413
+[#415]: https://github.com/qutip/QuantumToolbox.jl/issues/415