+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
+## [v0.28.0]
+Release date: 2025-02-22
+
 - Support for single `AbstractQuantumObject` in `sc_ops` for faster specific method in `ssesolve` and `smesolve`. ([#408])
 - Change save callbacks from `PresetTimeCallback` to `FunctionCallingCallback`. ([#410])
 - Align `eigenstates` and `eigenenergies` to QuTiP. ([#411])
@@ -115,6 +118,7 @@ Release date: 2024-11-13
 [v0.25.2]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.25.2
 [v0.26.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.26.0
 [v0.27.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.27.0
+[v0.28.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.28.0
 [#86]: https://github.com/qutip/QuantumToolbox.jl/issues/86
 [#139]: https://github.com/qutip/QuantumToolbox.jl/issues/139
 [#271]: https://github.com/qutip/QuantumToolbox.jl/issues/271